refactor(projection-art): Underwater demo scene 구조 분리
